requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
--find-links https://data.pyg.org/whl/torch-2.4.0+cu124.html

torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
torch-scatter

psutil
lightning==2.2
h5py
yacs
trimesh
scikit-image
loguru
boto3
mesh2sdf
tetgen
pymeshlab
plyfile
einops
libigl
polyscope
potpourri3d
simple_parsing
arrgh
open3d
vtk
omegaconf
diffusers
scipy
pygltflib
